email

+++ src/main/java/com/example/product/service/CambioServImpl.java
@@ -3,22 +3,29 @@ package com.example.product.service;
 import com.example.product.dto.ResponseTasa;
 import com.example.product.product.Cambio;
 import com.example.product.repository.RepositoryCambio;
-import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 @Component
-@AllArgsConstructor
 public class CambioServImpl implements CambioService{
+    @Autowired
     private RepositoryCambio repository;
+
+    @Autowired
+    private WebClient webClient;
+
 
     @Override
     public List<ResponseTasa> findAll() {
-        List<Cambio> cambios = repository.findAll();
+            List<Cambio> cambios = repository.findAll();
 
-        return !cambios.isEmpty() ? cambios.stream().map(this::cambiosToResponse).collect(Collectors.toList()) : null;
-    }
+            return !cambios.isEmpty() ? cambios.stream().map(this::cambiosToResponse).collect(Collectors.toList()) : null;
+
+        }
 
     @Override
     public ResponseTasa getById(Integer id) {
@@ -31,21 +38,35 @@ public class CambioServImpl implements CambioService{
     }
 
     @Override
-    public ResponseTasa insert(ResponseTasa request) {
-        Cambio cambio =  repository.save(this.requestToCambio(request));
-        return this.cambiosToResponse(cambio);
+    public ResponseTasa insert(ResponseTasa request, String token) {
+        if (validateToken(token)) {
+            Cambio cambio = repository.save(this.requestToCambio(request));
+            return this.cambiosToResponse(cambio);
+        } throw new RuntimeException("no tines los permisos");
     }
 
     @Override
-    public void delete(Integer id) {
-        repository.deleteById(id);
+    public void delete(Integer id,String token) {
+        if (validateToken(token)) {
+
+            repository.deleteById(id);
+        }
     }
 
     @Override
-    public ResponseTasa updateCambio(ResponseTasa request) {
+    public ResponseTasa updateCambio(ResponseTasa request,String token) {
+        if (Objects.isNull(request.getId())){
+            throw new IllegalArgumentException("el id no puede ser null");
+        }        if (validateToken(token)) {
 
-       Cambio cambio =  repository.save(this.requestToCambio(request));
-        return this.cambiosToResponse(cambio);
+            Cambio cambioExist = repository.findById(request.getId()).orElse(null);
+            cambioExist.setTasa(request.getTasa());
+            cambioExist.setNombre(request.getNombre());
+
+            Cambio cambio = repository.save(cambioExist);
+
+            return this.cambiosToResponse(cambio);
+        } throw new RuntimeException("no tines los permisos");
     }
 
     ResponseTasa cambiosToResponse(Cambio cambio){
@@ -66,5 +87,12 @@ public class CambioServImpl implements CambioService{
 
         );
     }
-
+    public boolean validateToken(String token) {
+        return webClient.get()
+                .uri("http://localhost:8080/api/user/validate")
+                .header("Authorization", "Bearer " + token)
+                .retrieve()
+                .bodyToMono(Boolean.class)
+                .block();
+    }
 }
